+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#14](https://github.com/LaxarJS/laxar-angular-adapter/issues/14): directives: removed layout directive
+    + **BREAKING CHANGE:** see ticket for details
 - [#13](https://github.com/LaxarJS/laxar-react-adapter/issues/13): eslint: use shared configuration
 - [#12](https://github.com/LaxarJS/laxar-react-adapter/issues/12): additional eslint restrictions
 - [#11](https://github.com/LaxarJS/laxar-react-adapter/issues/11): added dist-target for UMD-library